+++ src/lib/api.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { getToken, logout } from './auth';
+import { getToken, logout } from '../utils/LocalStorage';
 
 const api = axios.create({
   baseURL: 'http://localhost:3000', // Ganti sesuai API-mu
