+++ src/models/AuthModel.js
@@ -1,4 +1,5 @@
 import api from "../lib/api";
+import { getToken } from "../utils/LocalStorage";
 export class AuthModel {
   async login(email, password) {
     const res = await api.post("/login", { email, password });
@@ -7,6 +8,15 @@ export class AuthModel {
 
   async register(data) {
     const res = await api.post("/register", data);
+    return res.data;
+  }
+
+  async logout() {
+    const res = await api.post("/logout", {
+      headers: {
+        'Authorization': `Bearer ${getToken()}`,
+      }
+    });
     return res.data;
   }
 }
