+++ src/models/BookingModel.js
@@ -1,5 +1,5 @@
 import api from "../lib/api";
-import { getToken } from '../lib/auth';
+import { getToken } from '../utils/LocalStorage';
 export class BookingModel {
     async createBooking(data){
         const res = await api.post('/bookings',data,{
