+++ src/models/ChatModel.js
@@ -1,5 +1,5 @@
 import api from "../lib/api";
-import { getToken } from '../lib/auth';
+import { getToken } from '../utils/LocalStorage';
 export class ChatModel {
     async getMessages(chat_id) {
         const res = await api.get(`/chats/${chat_id}/messages`);
